requirements.txt
streamlit
torch
pillow
numpy
joblib
scikit-learn
